+++ breeze/config.py
@@ -29,11 +29,14 @@ class Config:
         :attr:`RECAPTCHA_PUBLIC_KEY`: A string containing the public key for use with Google's reCAPTCHA service.\n
         :attr:`RECAPTCHA_PRIVATE_KEY`: A string containing the private key for use with Google's reCAPTCHA service.
 
-    other configs::
+    oauth2 configs::
         :attr:`GITHUB_CLIENT_ID`: A string containing the client ID for use with GitHub's OAuth2 service.\n
         :attr:`GITHUB_CLIENT_SECRET`: A string containing the client secret for use with GitHub's OAuth2 service.\n
-        :attr:`GITHUB_SCOOP`: A string containing the scope for use with GitHub's OAuth2 service.
+        :attr:`GITHUB_SCOPE`: A string containing the scope for use with GitHub's OAuth2 service.
 
+        :attr:`DISCORD_CLIENT_ID`: A string containing the client ID for use with Discord's OAuth2 service.\n
+        :attr:`DISCORD_CLIENT_SECRET`: A string containing the client secret for use with Discord's OAuth2 service.\n
+        :attr:`DISCORD_SCOPE`: A string containing the scope for use with Discord's OAuth2 service.
 
     for more information on the configuration variables, see the documentation :doc:`/config` section.
 
@@ -72,11 +75,25 @@ class Config:
     GITHUB_CLIENT_ID = environ.get("GITHUB_CLIENT_ID", "")
     GITHUB_CLIENT_SECRET = environ.get("GITHUB_CLIENT_SECRET", "")
 
-    GITHUB_SCOOP = "user:email"  # we only need the user's email
+    GITHUB_SCOPE = "user:email"  # we only need the user's email
 
     if not (GITHUB_CLIENT_ID and GITHUB_CLIENT_SECRET):  # pragma: no cover
         logging.warning(
             "if you want to use github oauth2, please set GITHUB_CLIENT_ID and GITHUB_CLIENT_SECRET in your .env file"
             " and try again. Otherwise, you can ignore this warning. ( if you don't want to use github oauth2 or this) "
             "testing will continue without github oauth2."
+        )
+
+    # oauth2 with discord
+    DISCORD_CLIENT_ID = environ.get("DISCORD_CLIENT_ID", "")
+    DISCORD_CLIENT_SECRET = environ.get("DISCORD_CLIENT_SECRET", "")
+
+    # discord scope for get user name, email, and avatar
+    DISCORD_SCOPE = "identify email"
+
+    if not (DISCORD_CLIENT_ID and DISCORD_CLIENT_SECRET):  # pragma: no cover
+        logging.warning(
+            "if you want to use discord oauth2, please set DISCORD_CLIENT_ID and DISCORD_CLIENT_SECRET in your .env file"
+            " and try again. Otherwise, you can ignore this warning. ( if you don't want to use discord oauth2 or this) "
+            "testing will continue without discord oauth2."
         )